Initial Commit of code for Caesar Encryptor
I did mine in a different way than the lesson. I understand why the lesson chose to do it that way. I did not like the "repeat of alphabet" to accommodate a shift greater than 26 (number of letters in alphabet).

My solution is more eloquent in dealing with that but (my opinion). On the other hand, it could not be done in a simple, single function if statement. So, which is better? Your call. I wanted to complete the entire task using my version to ensure it worked. It may not be the best solution, but it works as required.