Initial Commit of Silent Auction code
Once again I've done things somewhat different than the instructor. This works fine and is nice and short. I originally had a function created to build the dictionary, but that seemed frivolous as there was only 1 line of code. The instructor uses a function for the determination of the winning bid. I'm not sure why. Seems to me that makes for additional coding since you only make this determination at the very end. My FOR loop works slick so I will just keep this whole program void of functions. Best practice?