Initial Commit of Program List and Flowchart
This is a tricky flowchart to diagram. I'm not sure I'm happy with it. I did not look at the instructor's version at any time so this is wholly my interpretation. It does what is required but I imagine there is a more eloquent way of documenting it. Notably without the circumference style arrows.  YMMV.