Initial Commit of Blackjack Code
Not the most eloquent solution. I'm quite positive of that. Probably rather ugly in some respects. But it appears to work (knock on wood).

I created this without reviewing any instructor hints. It is 100% my fumbling solution. I will next review the instructor solution in order to learn best methodology for this coding project. For now, though, a "successful" solution.