Initial Commit of code for Number Guessing Game
I was able to create this without using functions. That's probably missing the point as far as the instructor is concerned. I'm not sure what to think. My solution is perhaps not as technical rigorous but it's shorter and, on initial viewing, more obvious.

Still, I need to learn to incorporate these functions properly. I'm finding it difficult to wrap my head around creating functions for simple little tasks like assigning HARD vs EASY. Not intuitive to me at all when simple IF statements work just as well and quicker/more sensibly.