improvements based on teacher sampl
When I watched the teacher's solution, I found the problem with my original attempt to update the total_income variable from within a function.

I forgot I needed (or could) use the "global" qualifier for the variable.  Doing so makes it work.  So I reverted to my original attempt rather than the dictionary solution I resorted to.  This is better.

All in all, this isn't so bad by comparison with the teacher's solution.  My nested functions probably isn't the best but hey, it works and that's a hell of a start.