initial commit
First attempt at OOP.  I needed to refer to teacher's solution to get an grasp of what was needed to make these objects work.  I was on the right track, but also confused since the lessons didn't show "how" to utilize the modules properly.

I was also using the Class directly with the methods rather that creating an object(variable) first.  Not sure if this is just bad practice or what.  It worked but was not how the teacher did it.