initial commit
Well, I thought I'd do better on this to be honest.  And I did most of it ... eventually.  But there was a curveball tossed into the OOP regarding the car manager that stopped me in my tracks.  I hate when teacher's don't "teach" something prior to us being exposed to it.  Case in point, the class inheritance that occured in car manager where the class creation init didn't actually inherit anything but the car creation method inside the manager class DID.  Weird and was not something we had seen before.