initial commit
This wasn't much of an exercise from a length perspective.  I was able to create the dictionary myself using dictionary comprehension. I then created my list using the old loop method which was incorrect. I mean, it worked, but we were supposed to use list comprehension. D'uh.  Once I clued in to that, I was able to correctly use list comprehension to perform the task as well.  So, overall I'm getting this, just have a tendency to revert to old, clunkier ways.