initial commit
lots of instructor led coding today, but the final challenge was on the students.  I successfully completed the challenge AND did it exactly how the instructor ended up doing it in her reveal.  That was rewarding even if the challenge was relatively modest in scope.

I also tweaked my code to accommodate usage of different capitalization options by the user when entering websites, therefore eliminating misinformation when user searches on "facebook" when the database holds "Facebook"