initial commit
I attempted the hardest level of this challenge ... and succeeded!
As you might expect, I did almost nothing like the instructor's final code.  I used os module and had to learn some new stuff about pandas (& instead of AND).  It was a struggle, believe me.  I wish I could get this stuff to work right out of the shoot.  But, I'm happy to have succeeded without referring to any instructor hints or code.  That's a success even if my solution isn't as pretty as hers.