initial commit
Successfully completed this challenge and did so in short order for a change.  My solution was quite similar to the instructors, but not exact.  I can see where her use of additional functions smoothed things out a bit.  On the other hand, I like my version better because it won't send an email every 60 seconds all night long.  It sends one email once the ISS is in range and then stops.  If the ISS is not in range, then it continues checking every 60 seconds.  I suppose one could modify the code to send a second email once the ISS is no longer in range.  Regardless, I feel much better about this than the previous challenge which took me far too long to weed through.