initial commit
A lot of instructor guided coding again in this lesson.  And thank god!  There was some mind-wrecking new stuff out of the blue here.  Returning to classes and incorporating GUI was a huge leap from the previous few days.  I struggled.  Won't lie.  Some of the logic of passing arguments lost me.  Onwards and upwards.