initial commit
What a gongshow this turned out to be.  The lesson videos are obviously getting dated as some of the free services used here have changed their online interface.  Fun.  And then the whole mess with environmental variables in windows vs mac was a pain in the ass.  Thank heavens some previous students figured out the corrections and posted in comments.  Very frustrating lesson and a growing trend of late, I'm afraid.

This stuff was almost entirely instructor driven.  The few challenges were minimal and relatively simple.