initial commit
I successfully completed the extra hard challenge without referring to instructor's solutions. Now, I added a bunch of extra code that was unnecessary once I reviewed the teacher's solution. She incorporated List Comprehension to retrieve only the data she wanted from the API.  I, on the other hand, retrieved all data and then used code to determine which days of information to splice out.  It worked and is nifty, I suppose, but it took a long time and a lot of investigation to get working when it didn't even need to be done. List Comprehension looping allowed the first two entries to be plucked from the API data without a care in the world about actual dates.

Ah well, its still a solution and one I created independently.  That's a success in my book.