initial commit
THIS WAS A GARBAGE LESSON.  The challenge was to track exercise in a google sheet using nutritionix and sheety APIs. That was straight forward enough, but the instructions were often vague or had dated information compared to updated API documentation.  I admit I had to peruse solutions several times just to understand what the hell was being asked or required to make the API function properly.  Case in point, the data posting in Sheety requiring a nested dictionary ... what this actually meant was not explained well.  And then the whole replit environmental variable nonsense was irritating.  I saw the fixes other students came up with. I just didn't bother.  I do not like replit. SO, disappointing lesson all around.  Hopefully this is but an anomaly.