initial commit
I'll be honest, this was an extremely frustrating capstone. IMHO, it was not explained well. Parameters for the project were not revealed until the 4th Hint, which was not helpful to someone attempting the "hard" (i.e. no hints) version of the challenge.

Further more, there was not explanation of the instructor's solution which left me hanging on why all these classes were being created despite them not doing anything helpful imo. This was doubly frustrating since I was struggling with my own class creation and where to put __init__ methods. In many cases, they seemed unnecessary, but I didn't understand the implications of this.

Ultimately, I achieved a solution with only use of partial hints.  It was a mixed bag of success in that sense.  I required some help from hints, early on especially, but also feel my solution was derived, in part, on my own. It is different but not wildly so from the instructor's solution.  I even think a couple portions of my code are better. Others, not so much.

Should have been a better explained capstone though. And a more immersive solution explanation would have helped. I get not having too many hints and hand-holding at this stage, but the complete lack of a walkthrough was unappreciated.