initial commit
This is almost exclusively the instructors solution. I literally had zero interest in completing this portion of a horrible constructed capstone. Check Day 39 for my work in the first stage of the project.  This is just garbage.  it doesn't even work but I couldn't care less about debugging it. HORRIBLE CAPSTONE.