initial commit
Oh, boy, was this another awful experience.  The challenge itself isn't all that difficult. Or it wasn't in the end. But the documentation for spotipy and spotify (the former especially) was bloody awful. There is absolutely NO WAY you could come up with a solution by reading through the pertinent documentation from the tools used. Utter garbage as a lesson. Ultimately, the guts of the code was just copy pasted from other sources, students, instructor's solution.

I fail to see what that accomplished. It irritates me to no end that solutions are impossible to derive with the very documentation required to do such a task. One should not have to pilfer other's work online to make a simple piece of code buildable.  Not impressed.

I was able to do the scrape myself.  That's good at least.  The spotipy stuff I needed plenty of help but did do some bits myself once I had some guidance on WTF was going on.