initial commit
This was almost comically simple compared to Day 46's debacle.  I was able to solve this by myself. There were hints provided in the instructions. They were minor hints but they weren't exactly hidden, so it was impossible not to see them. So I think everyone doing this lesson they got a bit of help. Regardless, I was able to write the code solo and successfully debug. It works.