initial commit
An unfinished capstone. Once again, the target website has changed since the challenge was created. Spent plenty of time seeking solutions to the lazy loading issue. Only answer seems to be the script/json download option the other students discovered. I had capitulated and was ready to adopt this solution when Zillow stopped allowing my browser to connect to their website. Who knew they didn't like scraping?! Eye roll.

So, I'm done with this scraping portion of the course.  Too much dated challenges with websites that have changed specifically to deter scraping. If I ever need to do this kind of work with Selenium or BeautifulSoup, I'm confident I can bang my head to success. For the purposes of this python course, this section was too long and too quickly obsolete.