initial commit
This is admittedly a simple challenge though the whole decorator topic is a bit mind bending and not well explained in this course.  I was able to create a working solution by myself. I added a couple cosmetic changes once I viewed the solution (turned debug on and split return lines so not so long).