initial commit
The final challenge wasn't overly complicated, but I did have a hiccup midway through. A perusal of the instructor solution set off a lightbulb and i was able to complete my solution without further help. It is also a different solution than the instructor's.

I remain befuddled by the forced usage of classes in some of these challenges. I've either completely misunderstood something about oop or the instructor is simply forcing their use where it is wholly unnecessary. In fact, I'd go so far as to say the use of a class in this instance makes the code messier and more complicated than necessary. Not sure the benefit of it.