update with final challenge code
Well that went off the rails damned quick.
No idea how the instructor completed this. It works. I had not way of getting to a solution myself. Honestly, she puts links in as "hints" and they might as well just be kicks to the groin for all the value they provide.